pyproject.toml
[project]
name = "hurricane"
version = "0.3.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "aiogram>=3.18.0",
    "gitpython>=3.1.44",
    "kurigram>=2.1.39",
    "meval>=2.5",
    "qrcode>=8.0",
    "tgcrypto>=1.2.5",
]

[dependency-groups]
dev = [
    "ruff>=0.9.6",
]
